refactor: move unexpectedhttperrorlogger to testlab
