+++ packages/rest/src/__tests__/helpers.ts
@@ -8,29 +8,7 @@ import {
   RequestBodyObject,
   SchemaObject,
 } from '@loopback/openapi-v3-types';
-import {IncomingMessage} from 'http';
-import {LogError} from '..';
 import {RestServerConfig, RestServerResolvedConfig} from '../rest.server';
-
-export function createUnexpectedHttpErrorLogger(
-  expectedStatusCode: number = 0,
-): LogError {
-  return function logUnexpectedHttpError(
-    err: Error,
-    statusCode: number,
-    req: IncomingMessage,
-  ) {
-    if (statusCode === expectedStatusCode) return;
-
-    console.error(
-      'Unhandled error in %s %s: %s %s',
-      req.method,
-      req.url,
-      statusCode,
-      err.stack || err,
-    );
-  };
-}
 
 /**
  * Create an OpenAPI request body spec with the given content
