+++ packages/rest/src/__tests__/integration/http-handler.integration.ts
@@ -7,7 +7,12 @@ import {Context} from '@loopback/context';
 import {anOpenApiSpec, anOperationSpec} from '@loopback/openapi-spec-builder';
 import {ControllerSpec, get} from '@loopback/openapi-v3';
 import {ParameterObject, RequestBodyObject} from '@loopback/openapi-v3-types';
-import {Client, createClientForHandler, expect} from '@loopback/testlab';
+import {
+  Client,
+  createClientForHandler,
+  createUnexpectedHttpErrorLogger,
+  expect,
+} from '@loopback/testlab';
 import * as express from 'express';
 import * as HttpErrors from 'http-errors';
 import {is} from 'type-is';
@@ -30,7 +35,7 @@ import {
   UrlEncodedBodyParser,
   writeResultToResponse,
 } from '../..';
-import {createUnexpectedHttpErrorLogger, aRestServerConfig} from '../helpers';
+import {aRestServerConfig} from '../helpers';
 
 const SequenceActions = RestBindings.SequenceActions;
